Move Observer to layouts package
