+++ src/thumbnail/ThumbnailItem.tsx
@@ -8,8 +8,8 @@
 
 import React from 'react';
 
-import PdfJs from '../PdfJs';
 import Spinner from '../components/Spinner';
+import PdfJs from '../PdfJs';
 
 interface ThumbnailItemProps {
     page: PdfJs.Page;
